Fixed code generator; it no longer generates unnecessary casts for round and signum.

+++ trunk/ots-core/src/main/java/org/opentrafficsim/core/value/vfloat/FloatMathFunctionsImpl.java
@@ -92,7 +92,7 @@ public final class FloatMathFunctionsImpl
         @Override
         public float apply(final float a)
         {
-            return (float) Math.round(a);
+            return Math.round(a);
         }
     };
 
@@ -104,7 +104,7 @@ public final class FloatMathFunctionsImpl
         @Override
         public float apply(final float a)
         {
-            return (float) Math.signum(a);
+            return Math.signum(a);
         }
     };
